+++ src/id_translation/dio/_pandas.py
@@ -1,8 +1,9 @@
 import warnings
 from collections import defaultdict
-from typing import Any, Dict, Iterable, List, Optional, Sequence, TypeVar, Union
+from typing import Any, Dict, List, Optional, Sequence, TypeVar, Union
 
 import pandas as pd
+from pandas.api.types import is_float_dtype, is_numeric_dtype
 
 from ..offline import TranslationMap
 from ..types import IdType, NameType, SourceType
@@ -10,11 +11,8 @@ from ._data_structure_io import DataStructureIO
 from ._sequence import SequenceIO, translate_sequence, verify_names
 from .exceptions import NotInplaceTranslatableError
 
-T = TypeVar("T", pd.DataFrame, pd.Series)
-
-
-def _cast_series(series: pd.Series) -> pd.Series:
-    return series.dropna().astype(int) if series.dtype == float else series
+T = TypeVar("T", pd.DataFrame, pd.Series, pd.Index, pd.MultiIndex)
+PandasVectorT = TypeVar("PandasVectorT", pd.Series, pd.Index)
 
 
 class PandasIO(DataStructureIO):
@@ -41,64 +39,102 @@ class PandasIO(DataStructureIO):
             ans = defaultdict(list)
             for i, name in enumerate(translatable.columns):
                 if name in names:
-                    ans[name].extend(_cast_series(translatable.iloc[:, i]))
+                    ans[name].extend(_float_to_int(translatable.iloc[:, i]))
             return dict(ans)
-        elif isinstance(translatable, pd.Series):
-            return SequenceIO.extract(_cast_series(translatable), names)
         elif isinstance(translatable, pd.MultiIndex):
             # This will error on duplicate names, which is probably a good thing.
             return {name: list(translatable.unique(name)) for name in names}
-        elif isinstance(translatable, pd.Index):
-            return SequenceIO.extract(translatable, names)
+        elif isinstance(translatable, (pd.Series, pd.Index)):
+            verify_names(len(translatable), names)
+            translatable = _float_to_int(translatable)
+            if len(names) == 1:
+                return {names[0]: translatable.unique().tolist()}
+            else:
+                return SequenceIO.extract(translatable, names)
 
-        raise TypeError(f"This should not happen: {type(translatable)=}")
+        raise TypeError(f"This should not happen: {type(translatable)=}")  # pragma: no cover
 
     @staticmethod
     def insert(
         translatable: T, names: List[NameType], tmap: TranslationMap[NameType, SourceType, IdType], copy: bool
     ) -> Optional[T]:
-        if isinstance(translatable, (pd.DataFrame, pd.Series)):
-            if copy:
-                translatable = translatable.copy()
-        else:  # Index-type translatable
-            if not copy:
-                raise NotInplaceTranslatableError(translatable)  # TODO(issues/170): PDEP-6 check
+        if not copy:
+            if isinstance(translatable, pd.DataFrame):
+                pass  # Ok
+            elif isinstance(translatable, pd.Series):
+                pass  # Ok, for now.   # TODO(issues/170): PDEP-6 check
+            else:
+                raise NotInplaceTranslatableError(translatable)
+
+        if isinstance(translatable, pd.MultiIndex):
+            df = translatable.to_frame()
+            PandasIO.insert(df, names, tmap, copy=False)
+            return pd.MultiIndex.from_frame(df, names=translatable.names)
+
+        if isinstance(translatable, pd.Index):
+            result = _translate_pandas_vector(translatable, names, tmap)
+            if isinstance(result, pd.Index):
+                return result
+            return pd.Index(result, name=translatable.name, copy=False)
 
         if isinstance(translatable, pd.DataFrame):
+            if copy:
+                translatable = translatable.copy()
+
             for i, name in enumerate(translatable.columns):
                 if name in names:
-                    translatable.iloc[:, i] = _translate_series(translatable.iloc[:, i], [name], tmap)
-        elif isinstance(translatable, pd.Series):
-            verify_names(len(translatable), names)
-            result = _translate_series(translatable, names, tmap)
+                    translatable.iloc[:, i] = _translate_pandas_vector(translatable.iloc[:, i], [name], tmap)
+
+            return translatable if copy else None
+
+        if isinstance(translatable, pd.Series):
+            result = _translate_pandas_vector(translatable, names, tmap)
+
+            if copy:
+                if isinstance(result, pd.Series):
+                    return result
+                return pd.Series(result, index=translatable.index, name=translatable.name, copy=False)
+
             with warnings.catch_warnings():
                 warnings.simplefilter(action="ignore", category=FutureWarning)  # TODO(issues/170): PDEP-6 check
                 translatable[:] = result
-        elif isinstance(translatable, pd.MultiIndex):
-            df = translatable.to_frame()
-            PandasIO.insert(df, names, tmap, copy=False)
-            translatable = pd.MultiIndex.from_frame(df, names=translatable.names)
-        elif isinstance(translatable, pd.Index):
-            translatable = pd.Index(_translate_series(translatable, names, tmap), name=translatable.name)
-        else:
-            raise TypeError(f"This should not happen: {type(translatable)=}")
+            return None
 
-        return translatable if copy else None
+        raise TypeError(f"This should not happen: {type(translatable)=}")  # pragma: no cover
 
 
-def _translate_series(
-    # Not MultiIndex
-    series: Union[pd.Series, pd.Index],
+def _translate_pandas_vector(
+    pvt: PandasVectorT,
     names: List[NameType],
     tmap: TranslationMap[NameType, SourceType, IdType],
-) -> Iterable[Optional[str]]:
-    verify_names(len(series), names)
+) -> Union[List[Optional[str]], PandasVectorT]:
+    verify_names(len(pvt), names)
 
-    if len(names) == 1 and len(series) > 100:
-        # Optimization for large series. Suboptimal if "many" values are
-        # unique. Not worth the additional overhead for small series.
+    if len(names) == 1:
+        # Optimization for single-name vectors. Faster than SequenceIO for pretty much every size.
         magic_dict = tmap[names[0]]
-        mapping = {idx: magic_dict.get(idx) for idx in series.unique()}
-        return series.map(mapping)  # type: ignore[no-any-return]
+
+        mapping: Dict[IdType, Optional[str]]
+        if is_numeric_dtype(pvt):
+            # We don't need to cast float to int here, since hash(1.0) == hash(1). The cast in extract() is required
+            # because some database drivers may complain, especially if they receive floats (especially NaN).
+            mapping = {idx: magic_dict.get(idx) for idx in pvt.unique()}
+            return pvt.map(mapping)
+        else:
+            mapping = {}
+            data: List[Any] = pvt.to_list()
+            for i, idx in enumerate(data):
+                if idx in mapping:
+                    value = mapping[idx]
+                else:
+                    value = magic_dict.get(idx)
+                    mapping[idx] = value
+                data[i] = value
+
+        return data
     else:
-        return translate_sequence(series, names, tmap)
+        return translate_sequence(pvt, names, tmap)
+
+
+def _float_to_int(pvt: PandasVectorT) -> PandasVectorT:
+    return pvt.dropna().astype(int) if is_float_dtype(pvt) else pvt