+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.1] - 2024-03-09
+
 ### Changed
 - Heuristic functions that accept a `plural_to_singular`-argument now also accept a custom transformer.
 - Expose read-only attributes `Translator.fmt` and `default_fmt`.
@@ -207,7 +209,8 @@ cookiecutter template.
 - Switch to relative imports.
 - Fix some intersphinx issues.
 
-[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/rsundqvist/id-translator/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/rsundqvist/id-translator/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/rsundqvist/id-translator/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/rsundqvist/id-translator/compare/v0.5.0...v0.5.1